Port business logic to Variable Form Context
